requirements.txt
fastapi
uvicorn
chromadb
groq
python-dotenv
requests
pydantic
